unmapped_fast_api/requirements.txt
# UNMAPPED — Python dependencies
# Install: pip install -r requirements.txt

# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0

# ── Config validation ─────────────────────────────────────────────────────────
jsonschema==4.23.0

# ── AI — Gemini (Google) ──────────────────────────────────────────────────────
google-generativeai==0.8.0

# ── AI — Gemini ADK (talent + course agents) ─────────────────────────────────
# Optional: when installed, the /api/jobs/find and /api/courses/find routes
# perform live Google Search grounding via Gemini. When not installed, both
# routes transparently fall back to the curated lists in agent/adk/fallback_data.py.
google-adk==1.0.0

# ── Graph engine ──────────────────────────────────────────────────────────────
networkx==3.3

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
asyncpg==0.29.0
pgvector==0.3.2
alembic==1.13.2

# ── Data processing ───────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── QR code generation ────────────────────────────────────────────────────────
qrcode[pil]==7.4.2
Pillow==10.4.0

# ── PDF generation ────────────────────────────────────────────────────────────
weasyprint==62.1

# ── HTTP client (for external data APIs) ─────────────────────────────────────
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0        # async test client for FastAPI

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
